test_mongodb_atlas.py
#!/usr/bin/env python3
"""
Test script to verify MongoDB Atlas connection and all enhanced features
"""

import requests
import json
import time

def test_mongodb_atlas_connection():
    """Test MongoDB Atlas connection and features"""
    print("🧪 Testing MongoDB Atlas Connection & Enhanced Features")
    print("=" * 60)
    
    base_url = "http://localhost:5001"
    
    # Test 1: API Connection
    print("1. Testing API Connection...")
    try:
        response = requests.get(f"{base_url}/api/get_cakes", timeout=10)
        if response.status_code == 200:
            cakes = response.json()
            print(f"   ✅ API Connected - Found {len(cakes)} cakes")
            
            # Show sample cakes
            print("   📋 Sample Cakes:")
            for i, cake in enumerate(cakes[:3]):
                print(f"      {i+1}. {cake['name']} - {cake['category']} - Shs {cake['price']:,.0f}")
        else:
            print(f"   ❌ API Error: {response.status_code}")
            return False
    except Exception as e:
        print(f"   ❌ Connection Error: {e}")
        return False
    
    # Test 2: Homepage with Real Data
    print("\n2. Testing Homepage with Real Data...")
    try:
        response = requests.get(f"{base_url}/", timeout=10)
        if response.status_code == 200:
            if "Featured Cakes" in response.text:
                print("   ✅ Homepage loads with real data")
            else:
                print("   ⚠️  Homepage loaded but may not show real data")
        else:
            print(f"   ❌ Homepage Error: {response.status_code}")
    except Exception as e:
        print(f"   ❌ Homepage Error: {e}")
    
    # Test 3: Enhanced Wedding Gallery
    print("\n3. Testing Enhanced Wedding Gallery...")
    try:
        response = requests.get(f"{base_url}/wedding-cakes", timeout=10)
        if response.status_code == 200:
            if "Wedding & Special Occasions" in response.text:
                print("   ✅ Enhanced wedding gallery loaded")
            if "filter-tabs" in response.text:
                print("   ✅ Dynamic filtering features present")
            if "gallery-stats" in response.text:
                print("   ✅ Statistics section present")
            if "testimonials-section" in response.text:
                print("   ✅ Customer testimonials present")
        else:
            print(f"   ❌ Wedding Gallery Error: {response.status_code}")
    except Exception as e:
        print(f"   ❌ Wedding Gallery Error: {e}")
    
    # Test 4: Cake Details with Related Cakes
    print("\n4. Testing Cake Details with Related Cakes...")
    try:
        # Get first cake ID
        response = requests.get(f"{base_url}/api/get_cakes", timeout=10)
        if response.status_code == 200:
            cakes = response.json()
            if cakes:
                cake_id = cakes[0]['_id']
                response = requests.get(f"{base_url}/cake/{cake_id}", timeout=10)
                if response.status_code == 200:
                    if "More" in response.text and "Cakes" in response.text:
                        print("   ✅ Related cakes section present")
                    if "quick-actions" in response.text:
                        print("   ✅ Quick view/add to cart features present")
                    if "feature-tag" in response.text:
                        print("   ✅ Enhanced product details present")
                else:
                    print(f"   ❌ Cake Details Error: {response.status_code}")
    except Exception as e:
        print(f"   ❌ Cake Details Error: {e}")
    
    # Test 5: Category Filtering
    print("\n5. Testing Category Filtering...")
    try:
        categories = set(cake['category'] for cake in cakes)
        for category in list(categories)[:2]:  # Test first 2 categories
            response = requests.get(f"{base_url}/customer?category={category}", timeout=10)
            if response.status_code == 200:
                print(f"   ✅ Category '{category}' filtering works")
            else:
                print(f"   ❌ Category '{category}' filtering failed")
    except Exception as e:
        print(f"   ❌ Category Filtering Error: {e}")
    
    # Test 6: Database Collections
    print("\n6. Testing Database Collections...")
    try:
        # Test if we can access different collections
        response = requests.get(f"{base_url}/api/get_cakes", timeout=10)
        if response.status_code == 200:
            print("   ✅ Cakes collection accessible")
        
        # Test orders endpoint if it exists
        response = requests.get(f"{base_url}/orders", timeout=5)
        if response.status_code in [200, 302, 401]:  # 302/401 means endpoint exists but may require auth
            print("   ✅ Orders collection accessible")
    except Exception as e:
        print(f"   ❌ Database Collections Error: {e}")
    
    print("\n" + "=" * 60)
    print("🎉 MongoDB Atlas Integration Test Complete!")
    print("✅ Your FynCakes website is now running with:")
    print("   • Real MongoDB Atlas database connection")
    print("   • Enhanced wedding gallery with filtering")
    print("   • Related cakes on detail pages")
    print("   • Dynamic product information")
    print("   • Professional UI/UX improvements")
    print(f"\n🌐 Website URL: {base_url}")
    print("📱 Open in your browser to see all the enhancements!")

if __name__ == "__main__":
    test_mongodb_atlas_connection()
