requirements.txt
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.2
fonttools==4.59.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pymongo==4.14.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
Werkzeug==3.1.3
WTForms==3.2.1
